requirements.txt
Django==6.0.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
PyJWT==2.12.1
